requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.97.0
cryptography==41.0.3
python-dateutil==2.8.2
tqdm==4.66.1 